constant changes  - remove define and get_defined_constants  - all constants are persistent and can be discovered at compile time  - remove the fetch constant opcode (unused)  - remove case-insensitve constants (they're dumb)  - removed the need for the name variable in the constant structure (cause no more case insensitivity)  - increased the beginning size of the constants hash table up to 256
It's very likely I broke some parts of class constants with inheritence and stuff. there's a lot of stuff in the code I don't understand with the fetching of
variables and stuff.

There's a big nasty function called 'zval_update_constant' - which needs to be deleted... it's disgusting. Later, once I've revamped and removed magic